requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
llama-cpp-python>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
symspellpy>=6.7.0
numpy>=1.26.0
